refactor: centralize admin redirect helper
